requirements.txt
openai>=1.0.0
langchain>=0.0.300
langgraph>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.50.0
langchain-community>=0.3.29
langchain-openai>=0.3.33